pyproject.toml
[project]
name = "product-recommendation-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.8.0",
    "litellm>=1.74.8",
    "pydantic>=2.11.7",
]
